chat noti bug fix, recipe cleanup
