+++ src/views/RecipePage.vue
@@ -77,6 +77,7 @@ import ChipLabel from '@/components/Chip/ChipLabel.vue';
 import ParentHeader from '@/components/NavBar/ParentHeader.vue';
 import { mapGetters } from 'vuex';
 import { getResponse } from '@/helpers/getResponse';
+import { searchPrompts } from '@/data/searchPrompts';
 
 export default {
   name: 'RecipePage',
@@ -91,2777 +92,7 @@ export default {
     return {
       newTag: '',
       tags: [],
-      promptDICT: {
-        'Pantry Essentials': [
-          'butter',
-          'egg',
-          'garlic',
-          'milk',
-          'onion',
-          'sugar',
-          'flour',
-          'olive oil',
-          'garlic powder',
-          'rice',
-          'cinnamon',
-          'ketchup',
-          'soy sauce',
-          'mayonnaise',
-          'vegetable oil',
-          'bread',
-          'baking powder',
-          'brown sugar',
-          'oregano',
-          'potato',
-          'honey',
-          'paprika',
-          'baking soda',
-          'vanilla',
-          'spaghetti',
-          'peanut butter',
-          'chili powder',
-          'cumin',
-          'mustard',
-          'chicken breast',
-          'cheddar',
-          'onion powder',
-          'carrot',
-          'tomato',
-          'basil',
-          'parsley',
-          'parmesan',
-          'italian seasoning',
-          'thyme',
-          'bell pepper',
-        ],
-        'Vegetables & Greens': [
-          'garlic',
-          'onion',
-          'bell pepper',
-          'carrot',
-          'scallion',
-          'tomato',
-          'potato',
-          'red onion',
-          'celery',
-          'jalapeno',
-          'avocado',
-          'shallot',
-          'zucchini',
-          'cherry tomato',
-          'cucumber',
-          'spinach',
-          'sweet corn',
-          'chile pepper',
-          'sweet potato',
-          'broccoli',
-          'baby greens',
-          'pumpkin',
-          'cauliflower',
-          'cabbage',
-          'asparagus',
-          'kale',
-          'arugula',
-          'leek',
-          'eggplant',
-          'lettuce',
-          'butternut squash',
-          'romaine',
-          'beetroot',
-          'brussels sprout',
-          'fennel',
-          'sun dried tomato',
-          'radish',
-          'red cabbage',
-          'artichoke',
-          'new potato',
-          'summer squash',
-          'mixed greens',
-          'parsnip',
-          'baby carrot',
-          'poblano pepper',
-          'mixed vegetable',
-          'serrano pepper',
-          'sweet pepper',
-          'watercress',
-          'green tomato',
-          'iceberg',
-          'mashed potatoes',
-          'horseradish',
-          'chard',
-          'pimiento',
-          'butter lettuce',
-          'spaghetti squash',
-          'napa cabbage',
-          'celeriac',
-          'hash browns',
-          'water chestnut',
-          'turnip',
-          'thai chile pepper',
-          'bok choy',
-          'okra',
-          'acorn squash',
-          'radicchio',
-          'corn cob',
-          'pearl onion',
-          'tenderstem broccoli',
-          'plantain',
-          'baby bok choy',
-          'leaf lettuce',
-          'pepperoncini',
-          'jicama',
-          'endive',
-          'habanero pepper',
-          'corn husk',
-          'collard greens',
-          'french-fried onions',
-          'daikon',
-          'baby corn',
-          'broccoli rabe',
-          'belgian endive',
-          'rutabaga',
-          'ancho chile peppers',
-          'boston lettuce',
-          'yam',
-          'fresno chile',
-          'microgreens',
-          'kohlrabi',
-          'frisee',
-          'delicata squash',
-          'anaheim pepper',
-          'arbol chile pepper',
-          'cress',
-          'broccoli slaw',
-          'golden beet',
-          'pea shoots',
-          'gem lettuce',
-        ],
-        Mushrooms: [
-          'button mushroom',
-          'shiitake mushroom',
-          'portobello mushroom',
-          'porcini',
-          'wild mushroom',
-          'mixed mushrooms',
-          'oyster mushroom',
-          'chestnut mushroom',
-          'enoki mushroom',
-          'black fungus',
-          'black truffle',
-          'morel mushrooms',
-          'field mushroom',
-          'king oyster mushroom',
-          'shimeji mushroom',
-          'straw mushroom',
-          'dried chinese mushroom',
-          'maitake',
-          'trumpet mushroom',
-          'white truffle',
-          'white fungus',
-          'pioppini',
-          'snow fungus',
-          'white beech mushroom',
-          'boletus',
-          'huitlacoche',
-          'matsutake',
-          'nameko',
-          'djon djon mushroom',
-          'mixed asian mushrooms',
-          'puffball',
-          'honey fungus',
-          "caesar's mushroom",
-          'candy cap mushrooms',
-          'lion’s mane mushroom',
-        ],
-        Fruits: [
-          'lemon',
-          'lime',
-          'apple',
-          'orange',
-          'banana',
-          'raisins',
-          'pineapple',
-          'mango',
-          'peach',
-          'date',
-          'coconut',
-          'craisins',
-          'pear',
-          'grape',
-          'pomegranate',
-          'watermelon',
-          'rhubarb',
-          'dried apricot',
-          'kiwi',
-          'plum',
-          'grapefruit',
-          'fig',
-          'apricot',
-          'currant',
-          'mandarin',
-          'prunes',
-          'sultanas',
-          'cantaloupe',
-          'papaya',
-          'passion fruit',
-          'nectarine',
-          'tamarind',
-          'dried fig',
-          'chestnut',
-          'meyer lemon',
-          'honeydew melon',
-          'clementine',
-          'dried fruit',
-          'persimmon',
-          'melon',
-          'tangerine',
-          'dried mango',
-          'quince',
-          'dried apple',
-          'guava',
-          'kumquat',
-          'banana chips',
-          'jackfruit',
-          'dragon fruit',
-          'mixed fruit',
-          'asian pear',
-          'lychee',
-          'kaffir lime',
-          'star fruit',
-          'pomelo',
-          'tender coconut',
-          'chestnut puree',
-          'prickly pear',
-          'calamansi',
-          'yuzu',
-          'granadilla',
-          'apple chips',
-          'mixed peel',
-          'kokum',
-          'tangelo',
-          'dried lime',
-          'jujube',
-          'dried lemon',
-          'sweet lime',
-          'custard-apple',
-          'young jackfruit',
-          'durian',
-          'dried tamarind',
-          'freeze-dried apple',
-          'honey date',
-          'physalis',
-          'tamarillo',
-          'ice-apple',
-          'longan',
-          'finger lime',
-          'bitter orange',
-          'feijoa',
-          'dried persimmons',
-          'rambutan',
-          'rose apple',
-          'dried orange slice',
-          'loquat',
-          'crabapple',
-          'fig leaves',
-          'freeze-dried pineapple',
-          'soursop',
-          'pluot',
-          'hog plum',
-          'bergamot orange',
-          'luo han guo',
-          'mamey',
-          'sapote',
-          'green ume plum',
-          'kiwano',
-          'seagrape',
-        ],
-        Berries: [
-          'strawberry',
-          'blueberry',
-          'raspberry',
-          'cranberry',
-          'cherry',
-          'blackberry',
-          'berry mix',
-          'maraschino cherry',
-          'dried cherry',
-          'juniper berry',
-          'sour cherry',
-          'goji berry',
-          'dried blueberries',
-          'freeze-dried strawberry',
-          'gooseberry',
-          'freeze-dried raspberry',
-          'lingonberry',
-          'canned sour cherry',
-          'mulberry',
-          'acai berry',
-          'canned cherry',
-          'amla',
-          'elderberry',
-          'freeze-dried blueberry',
-          'huckleberry',
-          'dried elderberry',
-          'barberry',
-          'dried berries',
-          'sea buckthorn',
-          'saskatoon berry',
-          'rosehip',
-          'hawthorn',
-          'boysenberry',
-          'cloudberry',
-          'freeze-dried berries',
-          'aronia berry',
-          'chokeberry',
-          'loganberry',
-          'blackcurrant leaves',
-          'haskap berry',
-          'dewberry',
-          'sloe berry',
-          'oregon grape',
-        ],
-        'Nuts & Seeds': [
-          'walnut',
-          'pecan',
-          'almond',
-          'sesame seed',
-          'cashew',
-          'pine nut',
-          'pistachio',
-          'peanut',
-          'chia',
-          'flax',
-          'slivered almonds',
-          'pumpkin seeds',
-          'hazelnut',
-          'poppy seed',
-          'sunflower seeds',
-          'roasted peanuts',
-          'macadamia',
-          'chopped nuts',
-          'hemp hearts',
-          'nigella seeds',
-          'mixed nuts',
-          'brazil nut',
-          'mixed seeds',
-          'onion seed',
-          'watermelon seeds',
-          'honey-roasted peanuts',
-          'melon seeds',
-          'white chia',
-          'lotus seeds',
-          'trail mix',
-          'basil seeds',
-          'peanut brittle',
-          'candlenut',
-          'jackfruit seeds',
-          'honey-roasted almond',
-          'chironji',
-          'tigernuts',
-          'honey-roasted pecans',
-          'sunflower sprouts',
-          'apricot kernels',
-          'palm seeds',
-          'ginkgo nuts',
-          'keto trail mix',
-          'wattleseeds',
-          'barùkas',
-          'indian almond',
-        ],
-        Cheeses: [
-          'parmesan',
-          'cream cheese',
-          'cheddar',
-          'mozzarella',
-          'feta',
-          'ricotta',
-          'sharp cheddar',
-          'monterey jack',
-          'blue cheese',
-          'goat cheese',
-          'fresh mozzarella',
-          'swiss cheese',
-          'pecorino',
-          'gruyere',
-          'mascarpone',
-          'cottage cheese',
-          'american cheese',
-          'provolone',
-          'mexican cheese blend',
-          'pepper jack',
-          'brie',
-          'paneer',
-          'fontina',
-          'queso fresco',
-          'quark',
-          'gouda',
-          'cotija',
-          'asiago',
-          'smoked cheese',
-          'halloumi',
-          'chevre',
-          'manchego',
-          'italian cheese blend',
-          'neufchatel',
-          'herb cream cheese',
-          'burrata',
-          'havarti',
-          'colby',
-          'grana-padano',
-          'muenster',
-          'camembert',
-          'string cheese',
-          'soft cheese',
-          'stilton',
-          'raclette',
-          'cheddar-jack cheese',
-          'colby-jack cheese',
-          'taleggio',
-          'jarlsberg cheese',
-          'oaxaca',
-          'labneh',
-          'edam',
-          'creamy cheese wedge',
-          'cheese powder',
-          'fromage blanc',
-          'asadero',
-          'marble cheese',
-          'leicester',
-          'kefalotyri',
-          'lancashire',
-          'mizithra',
-          'kasseri',
-          'babybel',
-          'panela cheese',
-          'seasoned feta cheese',
-          'comté',
-          'longhorn',
-          'wensleydale',
-          'graviera',
-          'scamorza',
-          'cambozola',
-          'cheshire cheese',
-          'anthotyro',
-          'chenna',
-          'hard goat cheese',
-          'kashkaval',
-          'sheep cheese',
-          'amul cheese',
-          'reblochon',
-          'robiola',
-          'brick cheese',
-          'quick-melt cheese',
-          "farmer's cheese",
-          'mimolette',
-          'queso quesadilla',
-          'caciocavallo',
-          'manouri',
-          'requeijão',
-          'vacherin',
-          'brown cheese',
-          'gloucester',
-          'port salut',
-          'derby cheese',
-          'fontal',
-          'salad cheese',
-          'truffle cheese',
-          'epoisses cheese',
-          'maasdam',
-          'petit-suisse',
-          'sbrinz',
-        ],
-        'Dairy & Eggs': [
-          'butter',
-          'egg',
-          'milk',
-          'heavy cream',
-          'sour cream',
-          'buttermilk',
-          'yogurt',
-          'greek yogurt',
-          'cream',
-          'whipped cream',
-          'ghee',
-          'shortening',
-          'condensed milk',
-          'half and half',
-          'sweetened condensed milk',
-          'ice cream',
-          'margarine',
-          'creme fraiche',
-          'frosting',
-          'milk powder',
-          'curd',
-          'thickened cream',
-          'lemon curd',
-          'custard',
-          'dulce de leche',
-          'chocolate frosting',
-          'kefir',
-          'chocolate milk',
-          'sherbet',
-          'liquid egg substitute',
-          'whey',
-          'hung curd',
-          'quail egg',
-          'buttermilk powder',
-          'frozen yogurt',
-          'khoya',
-          'milk cream',
-          'coffee creamer',
-          'clotted cream',
-          'goat milk',
-          'cheese curd',
-          'sour milk',
-          'ganache',
-          'cajeta',
-          'duck egg',
-          'salted egg',
-          'skyr',
-          'raw milk',
-          'lime curd',
-          'powdered coffee creamer',
-          'chantilly',
-          'milkfat',
-          'yogurt starter',
-          'pumpkin spice coffee creamer',
-          'rainbow sherbet',
-          'strawberry frosting',
-          'amul butter',
-          'honey butter',
-          'strawberry cream cheese',
-          'goat butter',
-          'century egg',
-          'orange curd',
-          'goat yogurt',
-          'dahi',
-          'cinnamon sugar butter spread',
-          'bulgarian yogurt',
-          'tvorog',
-          'chocolate milk powder',
-          'liquid rennet',
-          'strawberry milk',
-          'ayran',
-          'cuajada',
-          'yogurt drink',
-          'pickled egg',
-          'sheep milk',
-          'starter culture',
-          'kashk',
-          'ostrich egg',
-          'yoplait whips',
-          'buffalo milk',
-          'goat kefir',
-          'lebneh',
-          'sheep’s milk yoghurt',
-          'vanilla milk',
-        ],
-        'Dairy-Free & Meat Substitutes': [
-          'coconut milk',
-          'almond milk',
-          'almond butter',
-          'tofu',
-          'coconut cream',
-          'vegan butter',
-          'non-dairy milk',
-          'soy milk',
-          'extra firm tofu',
-          'silken tofu',
-          'kala namak salt',
-          'coconut butter',
-          'egg replacer',
-          'vegan mayonnaise',
-          'vegan cheese',
-          'cashew butter',
-          'tempeh',
-          'vegan cream cheese',
-          'coconut yogurt',
-          'non-dairy yogurt',
-          'seed butter',
-          'cashew milk',
-          'oat milk',
-          'rice milk',
-          'nut butter',
-          'vegan sour cream',
-          'textured vegetable protein',
-          'soy yogurt',
-          'vegan worcestershire',
-          'vegan mozzarella',
-          'non-dairy creamer',
-          'vegan sausage',
-          'coconut whipped cream',
-          'smoked tofu',
-          'coconut powder',
-          'soy cream',
-          'seitan',
-          'coconut milk powder',
-          'non-dairy whipped topping',
-          'nut milk',
-          'non-dairy cream',
-          'vegan burger patty',
-          'condensed coconut milk',
-          'vegan ground beef',
-          'pulled oats',
-          'vegan bacon',
-          'soy curls',
-          'vegan pesto',
-          'marinated tofu',
-          'vegan feta',
-          'soy chorizo',
-          'hazelnut butter',
-          'hemp milk',
-          'vegan beef',
-          'vegan ranch',
-          'vegan chicken',
-          'vegetable suet',
-          'coconut paste',
-          'dairy-free ice-cream',
-          'almond-coconut milk',
-          'banana blossom',
-          'vegan fish sauce',
-          'vegetarian hot dog',
-          'hazelnut milk',
-          'maple almond butter',
-          'soy milk powder',
-          'vegan meatballs',
-          'almond-milk yogurt',
-          'almond creamer',
-          'vegan cream cheese frosting',
-          'coconut manna',
-          'falafel mix',
-          'ready-made falafel',
-          'vegan gravy',
-          'cashew cheese sauce',
-          'coconut fat',
-          'flax milk',
-          'hazelnut creamer',
-          'quorn',
-          'soy-free butter',
-          'tofurky',
-          'vegan nutella',
-          'vegan tzatziki',
-          'cashew cream cheese',
-          'cricket flour',
-          'macadamia butter',
-          'okara',
-          'egg tofu',
-          'protein drink',
-          'macadamia milk',
-          'vegan taco meat',
-          'walnut taco meat',
-          'vegan yogurt starter',
-          'banana milk',
-          'soy quark',
-          'vegan chicken nuggets',
-          'vegan starter culture',
-          'walnut milk',
-          'latik',
-          'rice cream',
-        ],
-        Meats: [
-          'bacon',
-          'ground beef',
-          'beef steak',
-          'ham',
-          'pork chops',
-          'sweet italian sausage',
-          'pork loin',
-          'prosciutto',
-          'sausage',
-          'beef roast',
-          'ground pork',
-          'beef stew meat',
-          'pepperoni',
-          'chorizo',
-          'pancetta',
-          'pork fillet',
-          'pork shoulder',
-          'ground lamb',
-          'pork ribs',
-          'smoked sausage',
-          'breakfast sausage',
-          'hot dog',
-          'beef sirloin',
-          'salami',
-          'brisket',
-          'deli ham',
-          'leg of lamb',
-          'beef short ribs',
-          'kielbasa',
-          'pork belly',
-          'andouille',
-          'boneless lamb',
-          'ground pork sausage',
-          'ground sausage',
-          'roast beef',
-          'bacon bits',
-          'pork roast',
-          'hot italian sausage',
-          'lamb shoulder',
-          'pork spare ribs',
-          'beef ribs',
-          'veal steak',
-          'lamb chops',
-          'bone-in ham',
-          'pork butt',
-          'canadian bacon',
-          'lamb shank',
-          'beef sausage',
-          'mutton',
-          'venison',
-          'bratwurst',
-          'ham steak',
-          'pulled pork',
-          'ham hock',
-          'frozen meatballs',
-          'rabbit',
-          'mixed ground meat',
-          'pork cutlets',
-          'veal cutlet',
-          'soup bones',
-          'pork steak',
-          'rack of lamb',
-          'lamb loin',
-          'mexican chorizo',
-          'pork back ribs',
-          'country style ribs',
-          'black forest ham',
-          'smoked ham hock',
-          'oxtail',
-          'serrano ham',
-          'raw chorizo',
-          'beef liver',
-          'pastrami',
-          'back bacon',
-          'cocktail sausage',
-          'hard salami',
-          'salt pork',
-          'veal shank',
-          'ground venison',
-          'blood sausage',
-          'beef shank',
-          'lap cheong',
-          'dried beef',
-          'gammon joint',
-          'country ham',
-          'boneless beef short ribs',
-          'boneless ham',
-          'ground bison',
-          'mortadella',
-          'fresh sausage',
-          'bologna',
-          'smoked pork chop',
-          'lamb neck',
-          'burger patty',
-          'sausage patty',
-          'beef suet',
-          'veal roast',
-          'beef shoulder',
-          'steak tips',
-          'veal chops',
-        ],
-        Poultry: [
-          'chicken breast',
-          'chicken thighs',
-          'cooked chicken',
-          'ground turkey',
-          'whole chicken',
-          'chicken leg',
-          'whole turkey',
-          'chicken wings',
-          'turkey breast',
-          'ground chicken',
-          'rotisserie chicken',
-          'chicken tenders',
-          'turkey sausage',
-          'chicken sausage',
-          'turkey bacon',
-          'duck',
-          'duck breast',
-          'boneless chicken',
-          'chicken livers',
-          'cornish hen',
-          'deli turkey',
-          'smoked turkey',
-          'turkey meat',
-          'chicken quarters',
-          'quail',
-          'ground turkey sausage',
-          'smoked turkey sausage',
-          'smoked chicken',
-          'turkey leg',
-          'pheasant',
-          'goose',
-          'turkey thigh',
-          'turkey pepperoni',
-          'turkey ham',
-          'chicken bones',
-          'foie gras',
-          'turkey meatballs',
-          'turkey neck',
-          'chicken giblets',
-          'turkey giblets',
-          'turkey wing',
-          'chicken nuggets',
-          'turkey burger',
-          'chicken andouille',
-          'chicken gizzards',
-          'smoked turkey leg',
-          'chicken italian sausage',
-          'crispy chicken strips',
-          'ostrich',
-          'popcorn chicken',
-          'chicken feet',
-          'pulled chicken',
-          'deli chicken',
-          'smoked duck breast',
-          'pigeon',
-          'turkey kielbasa',
-          'wild game bird',
-          'chicken necks',
-          'turkey liver',
-          'duck confit',
-          'roast duck',
-          'chicken meatballs',
-          'chicken-apple sausage',
-          'duck liver',
-          'guinea fowl',
-          'smoked turkey wings',
-          'chicken curry-cut',
-          'chicken schnitzel',
-          'grouse',
-          'chicken roast',
-          'goose liver',
-          'turkey bones',
-          'turkey lunch meat',
-          'turkey roast',
-          'capon',
-          'smoked turkey bacon',
-          'chicken bacon',
-          'turkey rissoles',
-          'chicken kebab',
-          'chicken ham',
-          'duck neck',
-          'chicken frames',
-          'duck bacon',
-          'chicken gyros',
-          'chicken patty',
-          'chicken ribs',
-          'turkey tail',
-          'chicken milanesa',
-        ],
-        Fish: [
-          'salmon',
-          'smoked salmon',
-          'cod',
-          'tilapia',
-          'tuna steak',
-          'whitefish',
-          'halibut',
-          'red snapper',
-          'sea bass',
-          'fish fillets',
-          'trout',
-          'catfish',
-          'swordfish',
-          'surimi',
-          'sardines',
-          'mahi mahi',
-          'sole',
-          'mackerel',
-          'smoked trout',
-          'caviar',
-          'haddock',
-          'smoked haddock',
-          'monkfish',
-          'flounder',
-          'perch',
-          'hake',
-          'salt cod',
-          'pollock',
-          'smoked mackerel',
-          'sea bream',
-          'rainbow trout',
-          'carp',
-          'cuttlefish',
-          'grouper',
-          'herring',
-          'salmon roe',
-          'steelhead trout',
-          'roe',
-          'barramundi',
-          'black cod',
-          'orange roughy',
-          'kingfish',
-          'turbot',
-          'bangus',
-          'rockfish',
-          'branzino',
-          'pomfret',
-          'arctic char',
-          'eel',
-          'dried anchovy',
-          'lemon sole',
-          'fresh anchovy',
-          'yellowtail',
-          'pickled herring',
-          'battered fish',
-          'pike',
-          'walleye',
-          'john dory',
-          'swai fish',
-          'fresh mackerel',
-          'salmon trout',
-          'basa fish',
-          'smoked eel',
-          'bluefish',
-          'fish balls',
-          'sturgeon',
-          'red mullet',
-          'gurnard',
-          'rohu fish',
-          'plaice',
-          'pompano',
-          'smoked fish',
-          'fish heads',
-          'dried fish',
-          'flathead',
-          'salt fish',
-          'smoked herring',
-          'whiting',
-          'fish cake',
-          'salmon burger meat',
-          'shark meat',
-          'garoupa',
-          'gilt-head bream',
-          'pangasius',
-          'salt herring',
-          'soused herring',
-          'tinapa',
-          'zander',
-          'amberjack',
-          'bottarga',
-          'korean fish cake',
-          'mullet',
-          'skipjack tuna',
-          'dried baby sardines',
-          'marlin',
-          'threadfin',
-          'tiny fish',
-          'tuna belly',
-          'beluga caviar',
-          'bombay duck',
-        ],
-        'Seafood & Seaweed': [
-          'shrimp',
-          'prawns',
-          'crab',
-          'scallop',
-          'mussel',
-          'clam',
-          'squid',
-          'nori',
-          'lobster',
-          'oyster',
-          'lobster tail',
-          'crawfish',
-          'octopus',
-          'kombu',
-          'dried shrimp',
-          'bay scallop',
-          'wakame',
-          'soft-shell crab',
-          'scampi',
-          'king crab',
-          'baby squid',
-          'mixed seafood',
-          'squid ink',
-          'dried prawns',
-          'dulse seaweed',
-          'roasted seaweed',
-          'smoked oyster',
-          'kelp',
-          'kizami nori',
-          'hijiki',
-          'salted shrimp',
-          'conch',
-          'yaki-nori',
-          'arame',
-          'mud crab',
-          'sea urchin',
-          'abalone',
-          'seaweed salad',
-          'dulse',
-          'smoked mussel',
-          'sea snails',
-          'aonori',
-          'prepared crab cake',
-          'sea lettuce',
-          'korean seaweed',
-          'ogo seaweed',
-          'seaweed caviar',
-          'shredded squid',
-          'sea moss',
-          'langoustine',
-        ],
-        'Herbs & Spices': [
-          'cinnamon',
-          'parsley',
-          'cilantro',
-          'cumin',
-          'basil',
-          'thyme',
-          'ginger root',
-          'garlic powder',
-          'oregano',
-          'nutmeg',
-          'chili flake',
-          'chili powder',
-          'paprika',
-          'cayenne',
-          'rosemary',
-          'bay leaf',
-          'turmeric',
-          'clove',
-          'onion powder',
-          'ginger powder',
-          'chive',
-          'dill',
-          'mint',
-          'green cardamom',
-          'smoked paprika',
-          'fresh mint',
-          'coriander powder',
-          'sage',
-          'coriander',
-          'allspice',
-          'cracked pepper',
-          'peppercorn',
-          'mustard seed',
-          'white pepper',
-          'curry leaves',
-          'fennel seed',
-          'tarragon',
-          'saffron',
-          'star anise',
-          'asafoetida',
-          'marjoram',
-          'lemongrass',
-          'caraway',
-          'celery seed',
-          'garlic granules',
-          'chipotle powder',
-          'chipotle',
-          'fenugreek',
-          'onion flake',
-          'matcha powder',
-          'ancho chile powder',
-          'sumac',
-          'dried parsley flakes',
-          'fenugreek seed',
-          'kashmiri red chilli',
-          'thai basil',
-          'edible flowers',
-          'aniseed',
-          'kaffir lime leaves',
-          'chervil',
-          'lavender',
-          'carom seeds',
-          'mace',
-          'mexican oregano',
-          'black mustard seed',
-          'mango powder',
-          'black cardamom',
-          'saffron strands',
-          'dried chilies',
-          'guajillo pepper',
-          'pink peppercorn',
-          'hot paprika',
-          'lemon thyme',
-          'galangal',
-          'garlic flakes',
-          'dried cilantro',
-          'lemon balm',
-          'dill seed',
-          'green peppercorn',
-          'aleppo pepper',
-          'wasabi powder',
-          'achiote seeds',
-          'savory herb',
-          'pandan leaves',
-          'sorrel',
-          'gochugaru',
-          'saigon cinnamon',
-          'lemongrass paste',
-          'shiso',
-          'celery powder',
-          'black cumin',
-          'vietnamese mint',
-          'anardana',
-          'dried orange peel',
-          'lemon verbena',
-          'espelette pepper',
-          'raw stevia',
-          'achiote paste',
-          'summer savory',
-          'fennel pollen',
-        ],
-        'Sugar & Sweeteners': [
-          'sugar',
-          'brown sugar',
-          'confectioners sugar',
-          'honey',
-          'maple syrup',
-          'corn syrup',
-          'coconut sugar',
-          'molasses',
-          'stevia',
-          'agave nectar',
-          'sugar syrup',
-          'erythritol',
-          'vanilla sugar',
-          'demerara sugar',
-          'caramel syrup',
-          'chocolate syrup',
-          'jaggery',
-          'raw sugar',
-          'golden syrup',
-          'cinnamon sugar',
-          'grenadine',
-          'liquid stevia',
-          'coarse sugar',
-          'salted caramel syrup',
-          'sanding sugar',
-          'dark corn syrup',
-          'sucralose',
-          'monk fruit sweetener',
-          'maple sugar',
-          'blackstrap molasses',
-          'glucose',
-          'rock sugar',
-          "confectioners' sweetener",
-          'xylitol',
-          'jam sugar',
-          'brown rice syrup',
-          'brown sugar substitute',
-          'strawberry syrup',
-          'vanilla syrup',
-          'ginger syrup',
-          'orgeat',
-          'rice malt syrup',
-          'pancake syrup',
-          'raspberry syrup',
-          'black treacle',
-          'date syrup',
-          'date paste',
-          'coconut syrup',
-          'mint syrup',
-          'treacle',
-          'rice syrup',
-          'manuka honey',
-          'maple butter',
-          'blueberry syrup',
-          'apple syrup',
-          'blackberry syrup',
-          'allulose',
-          'piloncillo',
-          'cherry syrup',
-          'hibiscus syrup',
-          'fresh sugar cane',
-          'lavender syrup',
-          'sorghum syrup',
-          'hazelnut syrup',
-          'white chocolate sauce',
-          'pumpkin spice syrup',
-          'glycerine',
-          'lucuma powder',
-          'cranberry syrup',
-          'black sugar',
-          'golden sugar',
-          'cane syrup',
-          'mango syrup',
-          'malt syrup',
-          'hot honey',
-          'gula melaka',
-          'elderberry syrup',
-          'dark chocolate syrup',
-          'inulin',
-          "sweet'n low",
-          'fructose',
-          'honey powder',
-          'grape syrup',
-          'berry syrup',
-          'brown butter syrup',
-          'date sugar',
-          'mastic gum',
-          'gum syrup',
-          'irish cream syrup',
-          'prickly pear syrup',
-          'cookie butter syrup',
-          'creamed honey',
-          'buttermilk syrup',
-          'chocolate sugar',
-          'flavored syrup',
-          'elderflower syrup',
-          'malt sugar',
-          'buckwheat honey',
-          'truffle honey',
-          'aspartame',
-        ],
-        'Seasonings & Spice Blends': [
-          'italian seasoning',
-          'seasoned salt',
-          'curry',
-          'garam masala',
-          'pumpkin pie spice',
-          'mustard powder',
-          'taco seasoning',
-          'cajun seasoning',
-          'dry ranch seasoning',
-          'white miso',
-          'himalayan salt',
-          'lemon & pepper seasoning',
-          'liquid smoke',
-          'poultry seasoning',
-          'chinese five spice',
-          'red curry',
-          'old bay seasoning',
-          'herbes de provence',
-          'chaat masala',
-          'creole seasoning',
-          'steak seasoning',
-          'mixed spice',
-          'fleur de sel',
-          'thai red curry paste',
-          'mixed herbs',
-          'green curry',
-          'barbecue seasoning',
-          'apple pie spice',
-          "za'atar",
-          'ras el hanout',
-          'seafood seasoning',
-          'all-purpose seasoning',
-          'bagel seasoning',
-          'fajita seasoning',
-          'tandoori spice',
-          'pickling spice',
-          'caribbean jerk seasoning',
-          'gravy mix',
-          'garlic-pepper seasoning',
-          'garlic & herb seasoning',
-          'pickling salt',
-          'greek seasoning',
-          'msg',
-          'teriyaki marinade',
-          'adobo seasoning',
-          'sambar powder',
-          'smoked salt',
-          'dash seasoning',
-          'red miso',
-          'hot curry',
-          'salt-free seasoning',
-          'yellow miso',
-          'carne asada seasoning',
-          'bouquet garni',
-          'dukkah',
-          'pav bhaji masala',
-          'curing salt',
-          'mexican seasoning',
-          'tajin seasoning',
-          'seasoned pepper',
-          'crab boil seasoning',
-          'shichimi togarashi',
-          'rib rub',
-          'truffle salt',
-          'biryani masala',
-          'furikake',
-          'togarashi',
-          'chipotle seasoning',
-          'meat masala',
-          'harissa spice blend',
-          'chicken gravy mix',
-          'grey sea salt',
-          'chana masala',
-          'au jus mix',
-          'chicken taco seasoning',
-          'mushroom seasoning',
-          'berbere',
-          'brown miso',
-          'gingerbread spice',
-          'popcorn seasoning',
-          'chili-lime seasoning',
-          'blackening seasoning',
-          'fines herbes',
-          'pizza seasoning',
-          'shawarma seasoning',
-          'chai masala',
-          'panang curry',
-          'kitchen king masala',
-          'lemon & herb seasoning',
-          'vanilla salt',
-          'thai seasoning',
-          'mulling spices',
-          'sriracha seasoning',
-          'hawaiian salt',
-          'coffee rub',
-          'mesquite seasoning',
-          'salad seasoning',
-          'tom yum paste',
-          'fish masala',
-          'fresh mixed herbs',
-        ],
-        Baking: [
-          'flour',
-          'vanilla',
-          'baking powder',
-          'baking soda',
-          'cornstarch',
-          'yeast',
-          'chocolate chips',
-          'dark chocolate chips',
-          'whole-wheat flour',
-          'almond flour',
-          'self-raising flour',
-          'shredded coconut',
-          'cornmeal',
-          'coconut flake',
-          'gelatin',
-          'pastry flour',
-          'cake flour',
-          'coconut flour',
-          'corn flour',
-          'baking mix',
-          'cream of tartar',
-          'bread flour',
-          'almond meal',
-          'unbleached all-purpose flour',
-          'white baking chips',
-          'gluten-free flour',
-          'rice flour',
-          'desiccated coconut',
-          'tapioca starch',
-          'yellow cake mix',
-          'chickpea flour',
-          'xanthan gum',
-          'oat flour',
-          'whole-wheat pastry flour',
-          'candy sprinkles',
-          'spelt',
-          'brownie mix',
-          'arrowroot flour',
-          'flaxseed meal',
-          'potato starch',
-          'chocolate cake mix',
-          'butterscotch chips',
-          'jello',
-          'peanut butter chips',
-          'crystallized ginger',
-          'buckwheat flour',
-          'brown rice flour',
-          'pectin',
-          'rye flour',
-          'psyllium husk',
-          'semolina flour',
-          'glutinous rice flour',
-          'pancake mix',
-          'muffin mix',
-          'refined flour',
-          'marzipan',
-          'coffee bean',
-          'toffee bits',
-          'cornbread mix',
-          'spice cake mix',
-          "devil's food cake mix",
-          'sorghum flour',
-          'masa harina',
-          'potato flakes',
-          'cinnamon chips',
-          'agar agar',
-          'red velvet cake mix',
-          'matzo meal',
-          'cassava flour',
-          'sago',
-          'white baking chocolate',
-          'sugar cookie mix',
-          'whipped cream stabilizer',
-          'vital wheat gluten',
-          'baking chips',
-          'meringue powder',
-          'hazelnut flour',
-          'citric acid',
-          'quinoa flour',
-          'coconut chips',
-          'finger millet flour',
-          'fondant',
-          'mint baking chips',
-          'angel food cake mix',
-          'white cornmeal',
-          'mincemeat',
-          'millet flour',
-          'gluten-free baking flour',
-          'einkorn flour',
-          'self-rising cornmeal',
-          'gluten-free self-raising flour',
-          'peanut flour',
-          'sweet bean paste',
-          'carob powder',
-          'tapioca pearls',
-          'guar gum',
-          'teff flour',
-          'potato flour',
-          'ready-made icing',
-          'seasoned flour',
-        ],
-        'Pre-Made Doughs & Wrappers': [
-          'pie crust',
-          'puff pastry',
-          'pizza crust',
-          'phyllo',
-          'refrigerated crescent rolls',
-          'biscuit dough',
-          'dumpling wrapper',
-          'rice paper',
-          'sourdough starter',
-          'cookie dough',
-          'egg roll wrapper',
-          'bread dough',
-          'butter puff pastry',
-          'frozen dinner roll',
-          'cinnamon roll dough',
-          'dosa batter',
-          'puff pastry shells',
-          'wontons',
-          'gyoza wrapper',
-          'wonton strips',
-          'gluten-free pizza crust',
-          'fresh pasta dough',
-          'idli batter',
-          'potsticker wrappers',
-          'chocolate pie crust',
-          'croissant dough',
-          'tofu skins',
-          'brisee',
-          'sausage rolls',
-          'kataifi',
-          'korean rice cake',
-          'corn dogs',
-          'tortilla dough',
-          'empanada wrapper',
-          'rough-puff pastry',
-          'gingerbread cookie dough',
-          'yufka',
-          'cob loaf',
-          'egg wrap',
-          'samosa sheet',
-          'frozen pizza rolls',
-          'brick pastry',
-        ],
-        'Grains & Cereals': [
-          'rolled oats',
-          'rice',
-          'quinoa',
-          'basmati rice',
-          'brown rice',
-          'quick-cooking oats',
-          'risotto rice',
-          'breakfast cereal',
-          'couscous',
-          'rice cereal',
-          'wild rice',
-          'semolina',
-          'polenta',
-          'jasmine rice',
-          'granola cereal',
-          'bulgur',
-          'pearl barley',
-          'farro',
-          'barley',
-          'wheat germ',
-          'grits',
-          'steel cut oats',
-          'millet',
-          'oat bran',
-          'glutinous rice',
-          'sushi rice',
-          'instant rice',
-          'hominy',
-          'red quinoa',
-          'raw buckwheat',
-          'wheat bran',
-          'wheat berries',
-          'instant oats',
-          'poha',
-          'black rice',
-          'yellow rice',
-          'fine semolina',
-          'muesli',
-          'amaranth',
-          'kasha',
-          'quinoa flakes',
-          'puffed rice',
-          'pearled farro',
-          'cracked wheat',
-          'freekeh',
-          'paella rice',
-          'sorghum',
-          'red rice',
-          'mexican rice',
-          'idli rice',
-          'cream of wheat',
-          'brown jasmine rice',
-          'rice pilaf',
-          'barnyard millet',
-          'weetabix',
-          'jambalaya rice mix',
-          'black glutinous rice',
-          'whole-grain oats',
-          'puffed quinoa',
-          'puffed wheat',
-          'cilantro lime rice',
-          'dirty rice mix',
-          'hulled barley',
-          'brown long grain rice',
-          'finger millet',
-          'foxtail millet',
-          'matta rice',
-          'rye berries',
-          'spelt flakes',
-          'kodo millet',
-          'oat groats',
-          'teff',
-          'rye flakes',
-          'little millet',
-          'gluten-free breakfast cereal',
-          'puffed amaranth',
-          'coconut rice',
-          'amaranth flakes',
-          'puffed kamut',
-          'bamboo rice',
-          'pearled sorghum',
-          'shirataki rice',
-          'vegetable fried rice',
-          'grünkern',
-          'jeera rice',
-          'fonio',
-          'kamut flakes',
-          'texmati rice',
-          'cream of rice',
-          'kañiwa',
-          'proso millet',
-        ],
-        Legumes: [
-          'peas',
-          'green beans',
-          'chickpea',
-          'black beans',
-          'kidney beans',
-          'white beans',
-          'lentils',
-          'pinto beans',
-          'snow peas',
-          'snap peas',
-          'red lentils',
-          'cannellini beans',
-          'bean sprouts',
-          'edamame',
-          'green lentils',
-          'urad dal',
-          'lima beans',
-          'chana dal',
-          'fava beans',
-          'black-eyed peas',
-          'moong dal',
-          'split peas',
-          'pigeon peas',
-          'red beans',
-          'mung bean sprout',
-          'soybeans',
-          'mung beans',
-          'string beans',
-          'yellow split peas',
-          'black lentils',
-          'borlotti beans',
-          'aquafaba',
-          'chana',
-          'snake beans',
-          'mixed beans',
-          'yellow lentils',
-          'fermented black beans',
-          'kala chana',
-          'black gram',
-          'wax beans',
-          'dried peas',
-          'pink beans',
-          'lentil sprouts',
-          'white pea',
-          'field peas',
-          'black soybeans',
-          'gigantes',
-          'flageolets',
-          'horse gram',
-          'soy sprouts',
-          'honey beans',
-          'cluster beans',
-          'brown beans',
-          'mayocoba beans',
-          'winged beans',
-          'castelluccio lentils',
-          'golden wax beans',
-          'moth beans',
-          'chickpea sprouts',
-          'hara chana',
-          'lupini beans',
-          'natto',
-          'hyacinth beans',
-          'petai',
-          'scarlet runner beans',
-          'soy flakes',
-        ],
-        Pasta: [
-          'short-cut pasta',
-          'spaghetti',
-          'macaroni',
-          'egg noodle',
-          'spiral pasta',
-          'lasagna',
-          'linguine',
-          'fettuccine',
-          'orzo',
-          'pasta shell',
-          'bow-tie pasta',
-          'noodle',
-          'tortellini',
-          'rice noodles',
-          'rigatoni',
-          'gnocchi',
-          'angel hair pasta',
-          'ramen noodles',
-          'vermicelli',
-          'tagliatelle',
-          'soba noodles',
-          'ravioli',
-          'ziti',
-          'orecchiette',
-          'israeli couscous',
-          'zoodles',
-          'udon noodles',
-          'ditalini',
-          'rice vermicelli',
-          'pappardelle',
-          'glass noodles',
-          'gluten-free pasta',
-          "mac 'n cheese",
-          'penne rigate',
-          'manicotti',
-          'bucatini',
-          'cannelloni',
-          'thai rice noodles',
-          'brown rice pasta',
-          'rotelle',
-          'shirataki noodles',
-          'chicken ramen',
-          'pierogi',
-          'soup pasta',
-          'sweet potato noodles',
-          'acini di pepe',
-          'cavatelli',
-          'instant noodle',
-          'somen noodles',
-          'yakisoba noodles',
-          'beef ravioli',
-          'hakka noodles',
-          'kelp noodles',
-          'fideo pasta',
-          'spaetzle',
-          'radiatore',
-          'paccheri',
-          'kluski noodles',
-          'matzo farfel',
-          'yolk-free noodles',
-          'black bean pasta',
-          'spinach fettuccine',
-          'rice-a-roni',
-          'fregola',
-          'frozen dumplings',
-          'beef tortellini',
-          'banh pho',
-          'butternut squash noodles',
-          'mie noodle',
-          'trahanas',
-          'busiate',
-          'lobster ravioli',
-          'mafalde',
-          'misua',
-          'palmini',
-          'trottole',
-          'mandu',
-          'pasta salad mix',
-          'pot pie noodle',
-          'casarecce',
-          'quinoa pasta',
-          'bean pasta',
-          'bean sheets',
-          'hilopites',
-          'jjolmyeon',
-          'naengmyeon noodles',
-          'yuba noodles',
-        ],
-        'Bread & Salty Snacks': [
-          'bread',
-          'bread crumbs',
-          'panko',
-          'flour tortillas',
-          'corn tortillas',
-          'crackers',
-          'baguette',
-          'tortilla chips',
-          'pita',
-          'pretzels',
-          'sourdough bread',
-          'rustic italian bread',
-          'popcorn',
-          'croutons',
-          'whole-wheat tortillas',
-          'english muffin',
-          'brioche',
-          'italian bread crumbs',
-          'rye bread',
-          'flatbread',
-          'dry-roasted peanuts',
-          'naan',
-          'potato chips',
-          'stuffing mix',
-          'cornbread',
-          'taco shells',
-          'tater tots',
-          'bagel',
-          'corn chips',
-          'unpopped popcorn',
-          'croissants',
-          'pork rinds',
-          'pumpernickel',
-          'muffin',
-          'sweet bread',
-          'pita chips',
-          'hawaiian rolls',
-          'potato bread',
-          'gluten free bread',
-          'breadsticks',
-          'focaccia',
-          'tostada shells',
-          'gluten-free bread crumbs',
-          'crispy fried onions',
-          'matzo',
-          'yeast extract spread',
-          'garlic bread',
-          'challah',
-          'roasted gram',
-          'crostini',
-          'rice cake',
-          'panettone',
-          'sweet potato fries',
-          'sev',
-          'vegetable chips',
-          'papad',
-          'spinach wraps',
-          'plantain chips',
-          'roasted chickpeas',
-          'cheetos',
-          'chapati',
-          'crumpet',
-          'seed bread',
-          'keto bread',
-          'bread bowl',
-          'sprouted grain bread',
-          'cornbread stuffing mix',
-          'low-carb wraps',
-          'frozen onion rings',
-          'corn nuts',
-          'corn muffin',
-          'pav bun',
-          'rusks',
-          'rice crackers',
-          'fruit bread',
-          'sprouted bread',
-          'pretzel bun',
-          'roti bread',
-          'crispbread',
-          'chocolate muffin',
-          'milk bread',
-          'popcorn shrimp',
-          'prawn crackers',
-          'wasabi peas',
-          'yorkshire pudding',
-          'bao bun',
-          'ezekiel bread',
-          'arabic bread',
-          'zwieback',
-          'boboli',
-          'oven chips',
-          'taco kit',
-          'gluten free pita',
-          'ready-made arepas',
-          'ready-made pancake',
-          'breading mix',
-          'japanese peanuts',
-          'khari boondi',
-          'potato waffle',
-          'tartlet shells',
-        ],
-        'Oils & Fats': [
-          'olive oil',
-          'vegetable oil',
-          'extra virgin olive oil',
-          'canola oil',
-          'coconut oil',
-          'cooking spray',
-          'sesame oil',
-          'frying oil',
-          'sunflower oil',
-          'avocado oil',
-          'toasted sesame oil',
-          'peanut oil',
-          'grapeseed oil',
-          'lard',
-          'corn oil',
-          'virgin coconut oil',
-          'chili oil',
-          'mustard oil',
-          'walnut oil',
-          'garlic oil',
-          'truffle oil',
-          'bacon grease',
-          'safflower oil',
-          'cacao butter',
-          'salad oil',
-          'duck fat',
-          'rice bran oil',
-          'soybean oil',
-          'butter-flavored cooking spray',
-          'white truffle oil',
-          'flaxseed oil',
-          'medium-chain triglyceride',
-          'pumpkin seed oil',
-          'hazelnut oil',
-          'coconut oil spray',
-          'almond oil',
-          'lemon oil',
-          'macadamia oil',
-          'goose fat',
-          'basil oil',
-          'palm oil',
-          'pork fat',
-          'beef dripping',
-          'schmaltz',
-          'tallow',
-          'sushi vinegar',
-          'pistachio oil',
-          'herb-infused olive oil',
-          'roasted peanut oil',
-          'argan oil',
-          'beef fat',
-          'pecan oil',
-          'crisco oil',
-          'red palm oil',
-          'shea butter',
-          'lamb fat',
-          'shallot oil',
-          'achiote oil',
-          'castor oil',
-          'chicken drippings',
-          'jojoba oil',
-          'oregano oil',
-          'hemp seed oil',
-          'wheat germ oil',
-          'ginger oil',
-          'cottonseed oil',
-          'pork dripping',
-          'rosehip seed oil',
-          'camelina oil',
-          'marula oil',
-          'moringa seed oil',
-          'white cacao butter',
-          'black seed oil',
-        ],
-        'Dressings & Vinegars': [
-          'mayonnaise',
-          'apple cider vinegar',
-          'balsamic vinegar',
-          'vinegar',
-          'red wine vinegar',
-          'rice wine vinegar',
-          'white wine vinegar',
-          'ranch dressing',
-          'italian dressing',
-          'sherry vinegar',
-          'distilled white vinegar',
-          'sweet chilli sauce',
-          'white balsamic vinegar',
-          'champagne vinegar',
-          'vinaigrette dressing',
-          'balsamic vinaigrette',
-          'blue cheese dressing',
-          'caesar dressing',
-          'thousand island',
-          'malt vinegar',
-          'black vinegar',
-          'canola mayonnaise',
-          'raspberry vinegar',
-          'japanese mayonnaise',
-          'tarragon vinegar',
-          'greek vinaigrette',
-          'sesame dressing',
-          'honey mustard dressing',
-          'aioli sauce',
-          'french dressing',
-          'spicy mayo',
-          'poppyseed dressing',
-          'catalina dressing',
-          'verjuice',
-          'cane vinegar',
-          'raspberry vinaigrette',
-          'red wine vinaigrette',
-          'russian dressing',
-          'avocado oil mayonnaise',
-          'champagne vinaigrette',
-          'sun-dried tomato vinaigrette',
-          'brown rice vinegar',
-          'garlic mayonnaise',
-          'coconut vinegar',
-          'ume plum vinegar',
-          'buttermilk ranch dressing',
-          'cilantro dressing',
-          'fig balsamic',
-          'spicy ranch dressing',
-          'green goddess dressing',
-          'creamy dressing',
-          'onion salad dressing',
-          'miso dressing',
-          'white balsamic vinaigrette',
-          'chardonnay vinegar',
-          'prepared coleslaw dressing',
-          'honey vinegar',
-          'tandoori mayonnaise',
-          'vinegar powder',
-          'chili vinegar',
-          'chili-lime dressing',
-          'moscatel vinegar',
-          'caramelised onion dressing',
-          'mango vinaigrette',
-          'sugarcane vinegar',
-          'spicy vinegar',
-          'creamy balsamic dressing',
-        ],
-        Condiments: [
-          'soy sauce',
-          'dijon mustard',
-          'worcestershire',
-          'hot sauce',
-          'ketchup',
-          'mustard',
-          'fish sauce',
-          'bbq sauce',
-          'sriracha',
-          'wholegrain mustard',
-          'tamari',
-          'oyster sauce',
-          'chili sauce',
-          'ginger-garlic paste',
-          'brown mustard',
-          'wing sauce',
-          'teriyaki sauce',
-          'prepared horseradish',
-          'ginger paste',
-          'dark soy sauce',
-          'coconut aminos',
-          'chili paste',
-          'harissa',
-          'cranberry sauce',
-          'chili-garlic sauce',
-          'tamarind paste',
-          'pomegranate molasses',
-          'gochujang',
-          'wasabi',
-          'honey mustard',
-          'mango chutney',
-          'english mustard',
-          'preserved lemon',
-          'sambal oelek',
-          'kecap manis',
-          'chutney',
-          'shrimp paste',
-          'picante sauce',
-          'thai sweet chili sauce',
-          'liquid aminos',
-          'sweet and sour sauce',
-          'tartar sauce',
-          'hot pepper jelly',
-          'chili-garlic paste',
-          'green chutney',
-          'bonito flakes',
-          'ponzu',
-          'mexican hot sauce',
-          'green chilli sauce',
-          'sambal',
-          'crispy onions',
-          'red pepper jelly',
-          'thai chili paste',
-          'peri peri',
-          'mint jelly',
-          'green chilli paste',
-          'horseradish mustard',
-          'aji amarillo',
-          'mustard paste',
-          'pepper paste',
-          'yuzu juice',
-          'tonkatsu sauce',
-          'tamarind chutney',
-          'chermoula',
-          'hp sauce',
-          'onion marmalade',
-          'duck sauce',
-          'ginger chili paste',
-          'giardiniera',
-          'doubanjiang',
-          'black soy sauce',
-          'korean bbq sauce',
-          'lime pickle',
-          'maggi sauce',
-          'chinese mustard',
-          'chamoy',
-          'chili crisp',
-          'doenjang',
-          'karashi',
-          'mushroom soy sauce',
-          'yuzu kosho',
-          'german mustard',
-          'taucheo',
-          'hot bean paste',
-          'remoulade',
-          'white bbq sauce',
-          'eel sauce',
-          'pickapeppa sauce',
-          'salsa lizano',
-          'patis',
-          'shrimp sauce',
-          'sweet soybean paste',
-          'grape must',
-          'shrimp powder',
-          'banana ketchup',
-          'chili puree',
-          'green harissa',
-          'crab paste',
-          'yellow bean paste',
-          'beer mustard',
-        ],
-        'Canned Food': [
-          'canned tomato',
-          'capers',
-          'green olives',
-          'canned chickpea',
-          'black olives',
-          'canned black beans',
-          'canned pumpkin',
-          'kalamata olives',
-          'canned tuna',
-          'pickle',
-          'canned pineapple',
-          'chipotle in adobo',
-          'canned anchovy',
-          'roasted red peppers',
-          'tomatoes with green chiles',
-          'canned artichoke',
-          'canned corn',
-          'canned kidney beans',
-          'canned pie filling',
-          'canned whole tomato',
-          'canned cannellini beans',
-          'refried beans',
-          'sweet pickle relish',
-          'sauerkraut',
-          'creamed corn',
-          'corned beef',
-          'canned beans',
-          'pickled jalapenos',
-          'diced green chiles',
-          'sun-dried tomatoes in oil',
-          'kimchi',
-          'canned mandarin oranges',
-          'chili beans',
-          'canned crab',
-          'bamboo shoot',
-          'baked beans',
-          'canned mushroom',
-          'canned salmon',
-          'pickling juice',
-          'dill pickle relish',
-          'canned peaches',
-          'canned black-eyed peas',
-          'pickled ginger',
-          'canned apple',
-          'canned green beans',
-          'spam',
-          'canned clam',
-          'chili with beans',
-          'pickled onions',
-          'fruit cocktail',
-          'canned lentils',
-          'canned peas',
-          'canned cherry tomato',
-          'pimiento-stuffed green olives',
-          'pickled red onion',
-          'canned pork',
-          'pickled beets',
-          'bread & butter pickles',
-          'canned apricot',
-          'canned pear',
-          'peppadew pepper',
-          'pickled vegetables',
-          'onion paste',
-          'canned baby corn',
-          'mexican-style corn',
-          'chili without beans',
-          'pork and beans',
-          'ranch-style beans',
-          'canned chicken breast',
-          'canned lychee',
-          'canned carrot',
-          'banana pepper rings',
-          'pumpkin pie filling',
-          'canned sardines',
-          'pickled pepper',
-          'tomato relish',
-          'canned jackfruit',
-          'taggiasca olives',
-          'canned sweet potato',
-          'cranberry relish',
-          'red pepper relish',
-          'canned asparagus',
-          'fire-roasted green chiles',
-          'corn relish',
-          'pickled hot pepper',
-          'golden hominy',
-          'canned mackerel',
-          'pickled cherry peppers',
-          'candied jalapenos',
-          'okra pickles',
-          'canned pimento',
-          'canned potato',
-          'umeboshi',
-          'tomato confit',
-          'brandied cherry',
-          'macapuno',
-          'canned four-bean mix',
-          'piccalilli',
-          'pickled green beans',
-          'half-sour pickle',
-        ],
-        'Sauces, Spreads & Dips': [
-          'peanut butter',
-          'tomato paste',
-          'tomato sauce',
-          'salsa',
-          'tahini',
-          'pesto',
-          'marinara sauce',
-          'pasta sauce',
-          'hoisin sauce',
-          'pico de gallo',
-          'stewed tomatoes',
-          'guacamole',
-          'hummus',
-          'enchilada sauce',
-          'fire-roasted tomato',
-          'salsa verde',
-          'alfredo sauce',
-          'balsamic glaze',
-          'anchovy paste',
-          'red enchilada sauce',
-          'steak sauce',
-          'chunky peanut butter',
-          'tzatziki',
-          'taco sauce',
-          'beef gravy',
-          'sun-dried tomato pesto',
-          'béchamel sauce',
-          'horseradish sauce',
-          'plum sauce',
-          'garlic butter',
-          'hollandaise sauce',
-          'black bean sauce',
-          'cocktail sauce',
-          'cheese dip',
-          'tapenade',
-          'turkey gravy',
-          'stir-fry sauce',
-          'thai peanut sauce',
-          'chicken gravy',
-          'sloppy joe sauce',
-          'pimento cheese spread',
-          'sofrito',
-          'mustard sauce',
-          'sun-dried tomato paste',
-          'demi-glace',
-          'queso dip',
-          'chimichurri sauce',
-          'french onion dip',
-          'herb butter',
-          'bolognese sauce',
-          'curry sauce',
-          'liver spread',
-          'browning sauce',
-          'schezwan sauce',
-          'mushroom gravy',
-          'pork gravy',
-          'country gravy',
-          'orange sauce',
-          'pineapple salsa',
-          'artichoke dip',
-          'japanese curry',
-          'black truffle butter',
-          'mango salsa',
-          'olive paste',
-          'spinach dip',
-          'roasted red pepper hummus',
-          'bacon jam',
-          'tikka masala sauce',
-          'vodka sauce',
-          'whipped cream cheese spread',
-          'kung pao sauce',
-          'onion gravy',
-          'sausage gravy',
-          'sesame sauce',
-          'mole paste',
-          'pad thai sauce',
-          'cheese spread',
-          'deviled ham spread',
-          'garlic spread',
-          'green olive tapenade',
-          'butter chicken sauce',
-          'black pepper sauce',
-          'char siu sauce',
-          'garlic pasta sauce',
-          'schezwan chutney',
-          'mentsuyu',
-          'black bean and corn salsa',
-          'tarama',
-          'ssamjang',
-          'thai red curry sauce',
-          'white pizza sauce',
-          'tomato tapenade',
-          'yum yum sauce',
-          'honey garlic sauce',
-          'mole sauce',
-          'nuoc cham',
-          'white clam sauce',
-          'bean dip',
-          'eggplant dip',
-          'pomegranate sauce',
-        ],
-        'Soups, Stews & Stocks': [
-          'chicken broth',
-          'vegetable broth',
-          'chicken stock',
-          'beef broth',
-          'beef stock',
-          'cream of mushroom',
-          'bouillon cube',
-          'cream of chicken',
-          'onion soup mix',
-          'tomato soup',
-          'fish stock',
-          'cream of celery',
-          'better than bouillon',
-          'clam juice',
-          'cheese soup',
-          'bone broth',
-          'dashi',
-          'onion soup',
-          'chicken soup',
-          'chicken bone broth',
-          'veal stock',
-          'turkey stock',
-          'lamb stock',
-          'turkey broth',
-          'beef consommé',
-          'stock paste',
-          'cream of potato',
-          'golden mushroom soup',
-          'seafood stock',
-          'mushroom broth',
-          'vegetable soup',
-          'chicken noodle soup',
-          'vegetable soup mix',
-          'chicken consommé',
-          'shrimp stock',
-          'nacho cheese soup',
-          'ramen noodle soup',
-          'cream of broccoli',
-          'lobster stock',
-          'pork stock',
-          'chicken rice soup',
-          'beefy onion soup mix',
-          'butternut squash soup',
-          'chicken stock paste',
-          'miso soup',
-          'broccoli-cheese soup',
-          'duck stock',
-          'clam broth',
-          'vegan chicken broth',
-          'cream of asparagus',
-          'ham stock',
-          'pork broth',
-          'beefy mushroom soup',
-          'new england clam chowder',
-          'bean soup mix',
-          'black bean soup',
-          'lentil soup',
-          'cream of shrimp soup',
-          'veal broth',
-          'vegetable beef soup',
-          'chicken soup mix',
-          'cream of bacon',
-          'lobster bisque',
-          'bean with bacon soup',
-          'clam chowder',
-          'condensed chicken gumbo soup',
-          'tortilla soup base',
-          'turkey bone broth',
-          'anchovy stock',
-          'cream of chicken soup mix:',
-          'minestrone',
-          'noodle soup mix',
-          'bovril',
-          'lamb broth',
-          'thai chicken broth',
-          'tomato bisque',
-          'venison stock',
-          'beef stock paste',
-          'chicken corn chowder',
-          'pork & beans',
-          'sun-dried tomato bisque mix',
-          'ham-flavored concentrate',
-          'oxtail soup',
-          'potato soup mix',
-          'callaloo',
-          'pho broth',
-          'split pea soup',
-          'chicken and mushroom soup',
-          'chili beef soup',
-          'corn chowder',
-          'cream of cauliflower',
-          'dashida',
-          'green pea soup',
-        ],
-        'Desserts & Sweet Snacks': [
-          'cocoa',
-          'dark chocolate',
-          'dark cocoa',
-          'chocolate',
-          'graham cracker',
-          'baking chocolate',
-          'marshmallow',
-          'applesauce',
-          'white chocolate',
-          'oreo',
-          'chocolate hazelnut spread',
-          'instant pudding',
-          'chocolate candy',
-          'dutch-process cocoa',
-          'raspberry jam',
-          'apricot jam',
-          'caramel sauce',
-          'candy coating',
-          'raw cacao powder',
-          'strawberry jam',
-          'biscuits',
-          'marshmallow creme',
-          'candy',
-          'jam',
-          'orange marmalade',
-          'wafer',
-          'cookies',
-          'peanut butter cups',
-          'chocolate pudding',
-          'candy cane',
-          'ginger snaps',
-          'cacao nibs',
-          'lady fingers',
-          'chocolate chip cookies',
-          'fudge sauce',
-          'digestive biscuits',
-          'chocolate cookies',
-          'apple butter',
-          'peppermint candy',
-          'cinnamon roll',
-          'butter cookies',
-          'candied cherry',
-          'vanilla pudding',
-          'caramel candies',
-          'currant jelly',
-          'candied ginger',
-          'angel food cake',
-          'peach preserves',
-          'chocolate wafer',
-          'candied peel',
-          'cherry jam',
-          'dark couverture chocolate',
-          'couverture chocolate',
-          'grape jelly',
-          'waffles',
-          'instant lemon pudding',
-          'fig jam',
-          'blueberry jam',
-          'cookie butter',
-          'butterscotch',
-          'candied fruit',
-          'almond cookies',
-          'blackberry preserves',
-          'apple jelly',
-          'gummies',
-          'candy corn',
-          'sugar cookies',
-          'ice-cream cone',
-          'marmalade',
-          'egg candies',
-          'strawberry puree',
-          'chocolate powder',
-          'sponge cake',
-          'chocolate-covered espresso bean',
-          'licorice',
-          'pineapple jam',
-          'plum jam',
-          'mexican chocolate',
-          'sorbet',
-          'white couverture',
-          'banana pudding',
-          'chocolate peanut butter',
-          'cinnamon candy',
-          'guava paste',
-          'pumpkin butter',
-          'fudge',
-          'strawberry sauce',
-          'butterscotch pudding mix',
-          'chocolate spread',
-          'doughnut',
-          'biscotti',
-          'raspberry sorbet',
-          'cheesecake instant pudding',
-          'peppermint patties',
-          'pistachio pudding',
-          'raspberry sauce',
-          'candied pineapple',
-          'chocolate fudge',
-          'hershey kisses',
-          'pistachio instant pudding',
-        ],
-        'Wine, Beer & Spirits': [
-          'white wine',
-          'red wine',
-          'whisky',
-          'rum',
-          'vodka',
-          'beer',
-          'orange liqueur',
-          'cider',
-          'tequila',
-          'sherry',
-          'gin',
-          'brandy',
-          'bitters',
-          'mirin',
-          'white rum',
-          'coffee liqueur',
-          'champagne',
-          'vermouth',
-          'irish cream',
-          'amaretto',
-          'marsala wine',
-          'cognac',
-          'sparkling wine',
-          'sake',
-          'rice wine',
-          'shaoxing wine',
-          'dry vermouth',
-          'liqueur',
-          'coconut rum',
-          'dessert wine',
-          'port wine',
-          'curacao',
-          'kirsch',
-          'peach schnapps',
-          'apple brandy',
-          'rosé wine',
-          'anise liqueur',
-          'herbal liqueur',
-          'limoncello',
-          'elderflower liqueur',
-          'cooking wine',
-          'hazelnut liqueur',
-          'peach liqueur',
-          'melon liqueur',
-          'raspberry liqueur',
-          'creme de cacao',
-          'madeira wine',
-          'schnapps',
-          'banana liqueur',
-          'absinthe',
-          'aperol',
-          'white cooking wine',
-          'vanilla vodka',
-          'cinnamon alcohol',
-          'creme de menthe',
-          'apricot brandy',
-          'cachaça',
-          'elderflower cordial',
-          'chocolate liqueur',
-          'ginger liqueur',
-          'sloe gin',
-          'maraschino',
-          'creme de cassis',
-          'cream sherry',
-          'bloody mary mix',
-          'gold rum',
-          'red cooking wine',
-          'sparkling rosé',
-          'grappa',
-          'lime cordial',
-          'mezcal',
-          'strawberry liqueur',
-          'drambuie',
-          'ginger wine',
-          'white port',
-          'pumpkin ale',
-          'bénédictine',
-          'peppermint liqueur',
-          'piña colada mix',
-          'advocaat',
-          'rumchata liqueur',
-          'coconut liqueur',
-          'daiquiri mix',
-          'galliano',
-          'blackberry brandy',
-          'plum wine',
-          'pisco',
-          'chocolate bitters',
-          'vanilla liqueur',
-          'sangria',
-          'grapefruit bitters',
-          'peach brandy',
-          'white chocolate liqueur',
-          'apple liqueur',
-          'pear brandy',
-          'rhum agricole',
-          'moonshine',
-          'armagnac',
-          'cherry vodka',
-          'bergamot liqueur',
-        ],
-        Beverages: [
-          'orange juice',
-          'coffee',
-          'club soda',
-          'espresso',
-          'pineapple juice',
-          'apple juice',
-          'tea',
-          'cranberry juice',
-          'coconut water',
-          'tomato juice',
-          'pomegranate juice',
-          'grapefruit juice',
-          'lemonade',
-          'coke',
-          'eggnog',
-          'ginger ale',
-          'ginger beer',
-          'orange juice concentrate',
-          'lemon lime soda',
-          'cream of coconut',
-          'sprite',
-          'green tea',
-          'lemonade concentrate',
-          'chai tea',
-          'root beer',
-          'tonic water',
-          'drinking chocolate',
-          'malted milk powder',
-          'mango juice',
-          'sour mix',
-          'hibiscus',
-          'tea leaves',
-          'grape juice',
-          'carrot juice',
-          'cherry juice',
-          'limeade concentrate',
-          'dr pepper',
-          'white grape juice',
-          'tangerine juice',
-          'passion-fruit juice',
-          'watermelon juice',
-          'fruit juice',
-          'iced tea',
-          'kombucha',
-          'apricot juice',
-          'beet juice',
-          'peach juice',
-          'orange soda',
-          'margarita mix',
-          'kool aid',
-          'energy drink',
-          'chamomile tea',
-          'pear juice',
-          'tamarind juice',
-          'tamarind water',
-          'cream soda',
-          'mountain dew',
-          'rooibos tea',
-          'grapefruit soda',
-          'lime soda',
-          'raspberry juice',
-          'guava juice',
-          'jasmine tea',
-          'clamato',
-          'strawberry juice',
-          'iced coffee concentrate',
-          'green tea leaves',
-          'beetroot juice',
-          'blueberry juice',
-          'lemonade mix',
-          'rose syrup',
-          'v8 juice',
-          'thai tea',
-          'aloe vera juice',
-          'white tea',
-          'juice blend',
-          'prune juice',
-          'sparkling cider',
-          'berry juice',
-          'butterfly pea flower',
-          'lapsang souchong',
-          'passion tea',
-          'strawberry soda',
-          'blackcurrant juice',
-          'herbal tea',
-          'banana juice',
-          'lychee juice',
-          'sugar cane juice',
-          'cranberry-raspberry juice',
-          'decaf coffee',
-          'pumpkin spice coffee',
-          'pepsi',
-          'peppermint tea',
-          'sports drink',
-          'acai berry juice',
-          'raspberry lemonade',
-          'lemon crystals',
-          'chicory coffee',
-          'kiwi juice',
-          'apple tea',
-        ],
-        'Supplements & Extracts': [
-          'almond extract',
-          'food coloring',
-          'nutritional yeast',
-          'peppermint extract',
-          'protein powder',
-          'lemon extract',
-          'coconut extract',
-          'rose water',
-          'orange extract',
-          'rum extract',
-          'maple extract',
-          'collagen',
-          'orange blossom water',
-          'chocolate protein powder',
-          'liquid egg white',
-          'peanut butter powder',
-          'vegan protein powder',
-          'essence',
-          'maca powder',
-          'spirulina',
-          'coffee extract',
-          'strawberry extract',
-          "brewer's yeast",
-          'butter extract',
-          'chocolate extract',
-          'raspberry extract',
-          'bee pollen',
-          'anise extract',
-          'cannabis',
-          'banana extract',
-          'lavender oil',
-          'essential oil',
-          'chicken essence',
-          'caramel extract',
-          'egg white powder',
-          'vitamin c',
-          'cannabutter',
-          'root beer extract',
-          'acai powder',
-          'hemp protein',
-          'ube flavoring',
-          'glucomannan',
-          'hazelnut extract',
-          'freeze-dried strawberry powder',
-          'tamarind extract',
-          'cherry extract',
-          'butterscotch flavor',
-          'kewra water',
-          'pineapple extract',
-          'lemon juice concentrate',
-          'chocolate collagen',
-          'cinnamon extract',
-          'cannabis milk',
-          'malt extract',
-          'kombucha starter',
-          'pandan extract',
-          'soy lecithin',
-          'camu powder',
-          'wheatgrass powder',
-          'ashwagandha',
-          'casein',
-          'cbd oil',
-          'chlorella',
-          'fish oil',
-          'probiotics',
-          'lime essential oil',
-          'activated charcoal',
-          'egg powder',
-          'reishi mushroom',
-          'vitamin e',
-          'wine yeast',
-          'barley grass',
-          'greens powder',
-          'rice protein powder',
-          'tea-tree oil',
-          'vitamin d',
-          'calcium lactate',
-          'magnesium',
-          'mango extract',
-          'raspberry powder',
-          'blueberry extract',
-          'corn extract',
-          'creatine',
-          'daily vitamins',
-          'moringa powder',
-          'pure lime extract',
-          'sodium alginate',
-          'sunflower lecithin',
-          'thc',
-          'berry powder',
-          'champagne yeast',
-          'maqui',
-          'rose oil',
-          'banana powder',
-          'chaga mushroom powder',
-          'clove oil',
-          'cranberry powder',
-          'eucalyptus oil',
-          'goji berry powder',
-          'maltodextrin',
-        ],
-      },
+      promptDICT: searchPrompts,
       prompt: {},
     };
   },
